requirements.txt
# ETL Pipeline Dependencies
pandas==2.1.4
numpy==1.24.3
sqlalchemy==2.0.23
sqlite3
pyyaml==6.0.1
requests==2.31.0
psycopg2-binary==2.9.9
great-expectations==0.18.8
apache-airflow==2.8.1
pytest==7.4.4
pytest-cov==4.1.0
logging
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
plotly==5.17.0
boto3==1.34.3
snowflake-connector-python==3.6.0
dbt-core==1.7.4
